server/requirements.txt
Flask
flask-bcrypt
python-dotenv
PyJWT
psycopg2-binary
pyotp
qrcode[pil]
jsonschema
requests
cryptography
gunicorn
geoip2==4.7.0